+++ testinfra/backend/local.py
@@ -1,4 +1,3 @@
-# coding: utf-8
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
@@ -11,8 +10,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from __future__ import unicode_literals
-
 from testinfra.backend import base
 
 
@@ -20,7 +17,7 @@ class LocalBackend(base.BaseBackend):
     NAME = "local"
 
     def __init__(self, *args, **kwargs):
-        super(LocalBackend, self).__init__("local", **kwargs)
+        super().__init__("local", **kwargs)
 
     def get_pytest_id(self):
         return "local"