+++ testinfra/modules/__init__.py
@@ -1,4 +1,3 @@
-# coding: utf-8
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
@@ -11,8 +10,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from __future__ import unicode_literals
-
 import importlib
 
 modules = {
@@ -20,6 +17,7 @@ modules = {
     'ansible': 'ansible:Ansible',
     'command': 'command:Command',
     'docker': 'docker:Docker',
+    'podman': 'podman:Podman',
     'environment': 'environment:Environment',
     'file': 'file:File',
     'group': 'group:Group',