+++ testinfra/modules/package.py
@@ -1,4 +1,3 @@
-# coding: utf-8
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
@@ -11,8 +10,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from __future__ import unicode_literals
-
 from testinfra.modules.base import Module
 
 
@@ -21,7 +18,7 @@ class Package(Module):
 
     def __init__(self, name):
         self.name = name
-        super(Package, self).__init__()
+        super().__init__()
 
     @property
     def is_installed(self):
@@ -64,7 +61,6 @@ class Package(Module):
         return "<package %s>" % (self.name,)
 
     @classmethod
-    # pylint: disable=too-many-return-statements
     def get_module_class(cls, host):
         if host.system_info.type == 'windows':
             return ChocolateyPackage