+++ testinfra/modules/salt.py
@@ -1,4 +1,3 @@
-# coding: utf-8
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
@@ -11,12 +10,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from __future__ import unicode_literals
-
-
 import json
-
-import six
 
 from testinfra.modules.base import InstanceModule
 
@@ -37,7 +31,7 @@ class Salt(InstanceModule):
 
     def __call__(self, function, args=None, local=False, config=None):
         args = args or []
-        if isinstance(args, six.string_types):
+        if isinstance(args, str):
             args = [args]
         if self._host.backend.HAS_RUN_SALT:
             return self._host.backend.run_salt(function, args)