+++ testinfra/plugin.py
@@ -1,4 +1,3 @@
-# coding: utf-8
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
@@ -10,9 +9,6 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-# pylint: disable=redefined-outer-name
-
-from __future__ import unicode_literals
 
 import logging
 import shutil
@@ -21,7 +17,6 @@ import tempfile
 import time
 
 import pytest
-import six
 import testinfra
 import testinfra.host
 import testinfra.modules
@@ -48,7 +43,7 @@ def pytest_addoption(parser):
         dest="connection",
         help=(
             "Remote connection backend (paramiko, ssh, safe-ssh, "
-            "salt, docker, ansible)"
+            "salt, docker, ansible, podman)"
         )
     )
     group.addoption(
@@ -127,7 +122,7 @@ def pytest_generate_tests(metafunc):
             "_testinfra_host", params, ids=ids, scope="module", indirect=True)
 
 
-class NagiosReporter(object):
+class NagiosReporter:
 
     def __init__(self, out):
         self.passed = 0
@@ -169,26 +164,18 @@ class NagiosReporter(object):
         return ret
 
 
-if six.PY2:
-    class SpooledTemporaryFile(tempfile.SpooledTemporaryFile):
+class SpooledTemporaryFile(tempfile.SpooledTemporaryFile):
 
-        def __init__(self, encoding=None, *args, **kwargs):
-            # tempfile.SpooledTemporaryFile is not new style class
-            tempfile.SpooledTemporaryFile.__init__(self, *args, **kwargs)
-            self.encoding = encoding
-else:
-    class SpooledTemporaryFile(tempfile.SpooledTemporaryFile):
+    def __init__(self, *args, **kwargs):
+        self._out_encoding = kwargs['encoding']
+        super().__init__(*args, **kwargs)
 
-        def __init__(self, *args, **kwargs):
-            self._out_encoding = kwargs['encoding']
-            super().__init__(*args, **kwargs)
-
-        def write(self, s):
-            # avoid traceback in py.io.terminalwriter.write_out
-            # TypeError: a bytes-like object is required, not 'str'
-            if isinstance(s, str):
-                s = s.encode(self._out_encoding)
-            return super().write(s)
+    def write(self, s):
+        # avoid traceback in py.io.terminalwriter.write_out
+        # TypeError: a bytes-like object is required, not 'str'
+        if isinstance(s, str):
+            s = s.encode(self._out_encoding)
+        return super().write(s)
 
 
 @pytest.mark.trylast